Fix issue with seeders
