+++ package.json
@@ -56,6 +56,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "faker": "^5.5.3",
+    "flatted": "^3.2.7",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "pg": "^8.5.1",
