+++ server/controllers/states.js
@@ -12,6 +12,63 @@ const { State, ApiLogs } = models;
  */
 class StateController {
   /**
+   * Get states
+   * @static
+   * @param {*} req - Request object
+   * @param {*} res - Response object
+   * @return {json} Returns json object
+   * @memberof StateController
+   */
+  static async getAll(req, res, next) {
+    const apilog = {
+      name: `${StateController.parameters.toLowerCase()}.getAll`,
+      refid: randString(`${StateController.parameter.toUpperCase()}`),
+      reqbody: JSON.stringify(req.body),
+      resbody: "",
+      httpstatuscode: 200,
+      statuscode: 200,
+      message: `${StateController.parameters} retrieved successfully`,
+      apiref: v4(),
+      url: `${req.method} ~ ${req.originalUrl}`,
+      reqstarttime: Date.now(),
+      reqendtime: "",
+    };
+
+    try {
+      const payload = await State.findAll({
+        order: [["name", "ASC"]]
+      });
+
+      apilog.resbody = JSON.stringify(payload);
+      apilog.reqendtime = Date.now();
+      await ApiLogs.create({ ...apilog });
+
+      return ResponseController.success(
+        res,
+        200,
+        200,
+        `${StateController.parameters} retrieved successfully`,
+        payload
+      );
+    } catch (err) {
+      apilog.resbody = JSON.stringify(err);
+      apilog.httpstatuscode = 400;
+      apilog.statuscode = 400;
+      apilog.message = `${StateController.parameters} could not be retrieved`;
+      apilog.reqendtime = Date.now();
+      await ApiLogs.create({ ...apilog });
+
+      return ResponseController.error(
+        res,
+        400,
+        400,
+        `${StateController.parameters} could not be retrieved`,
+        err
+      );
+    }
+  }
+
+  /**
    * Get a state item by id
    * @static
    * @param {*} req - Request object
