--- server/seeders/007-permissionroles.js
@@ -1,462 +1,0 @@
-module.exports = {
-  up: (queryInterface) => queryInterface.bulkInsert("PermissionRoles", [
-    {
-      roleid: 1,
-      permissionid: 1,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 2,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 3,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 4,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 5,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 6,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 7,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 8,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 9,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 10,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 11,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 12,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 13,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 14,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 15,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 16,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 17,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 18,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 19,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 20,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 21,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 22,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 23,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 24,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 25,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 26,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 27,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 28,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 29,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 30,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 31,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 32,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 33,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 34,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 35,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 36,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 37,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 38,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 39,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 40,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 41,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 42,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 43,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 44,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 45,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 46,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 47,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 48,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 49,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 50,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 51,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 52,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 53,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 54,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 55,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 56,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 57,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 58,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 59,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 60,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 61,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 62,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 63,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 64,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 65,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 66,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 67,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 68,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 69,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 70,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 71,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 72,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 73,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 74,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 75,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-    {
-      roleid: 1,
-      permissionid: 76,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    },
-  ], {}),
-
-  down: (queryInterface) => queryInterface.bulkDelete("PermissionRoles", null, {})
-};
