+++ server/seeders/009-permissions.js
@@ -1,0 +1,462 @@
+module.exports = {
+  up: (queryInterface) => queryInterface.bulkInsert("Permissions", [
+    {
+      name: "can:post:zone",
+      notes: "can post zone",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:zone",
+      notes: "can get zone",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:zone",
+      notes: "can put zone",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:zone",
+      notes: "can delete zone",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:branch",
+      notes: "can post branch",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:branch",
+      notes: "can get branch",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:branch",
+      notes: "can put branch",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:branch",
+      notes: "can delete branch",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:fellowship",
+      notes: "can post fellowship",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:fellowship",
+      notes: "can get fellowship",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:fellowship",
+      notes: "can put fellowship",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:fellowship",
+      notes: "can delete fellowship",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:event",
+      notes: "can post event",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:event",
+      notes: "can get event",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:event",
+      notes: "can put event",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:event",
+      notes: "can delete event",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:category",
+      notes: "can post category",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:category",
+      notes: "can get category",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:category",
+      notes: "can put category",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:category",
+      notes: "can delete category",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:preacher",
+      notes: "can post preacher",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:preacher",
+      notes: "can get preacher",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:preacher",
+      notes: "can put preacher",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:preacher",
+      notes: "can delete preacher",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:config",
+      notes: "can post config",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:config",
+      notes: "can get config",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:config",
+      notes: "can put config",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:config",
+      notes: "can delete config",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:user",
+      notes: "can post user",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:user",
+      notes: "can get user",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:user",
+      notes: "can put user",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:user",
+      notes: "can delete user",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:permission",
+      notes: "can post permission",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:permission",
+      notes: "can get permission",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:permission",
+      notes: "can put permission",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:permission",
+      notes: "can delete permission",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:role",
+      notes: "can post role",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:role",
+      notes: "can get role",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:role",
+      notes: "can put role",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:role",
+      notes: "can delete role",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:apilog",
+      notes: "can post apilog",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:apilog",
+      notes: "can get apilog",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:apilog",
+      notes: "can put apilog",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:apilog",
+      notes: "can delete apilog",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:download",
+      notes: "can post download",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:download",
+      notes: "can put download",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:download",
+      notes: "can get download",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:download",
+      notes: "can delete download",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:activitytype",
+      notes: "can post activitytype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:activitytype",
+      notes: "can put activitytype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:activitytype",
+      notes: "can get activitytype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:activitytype",
+      notes: "can delete activitytype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:trainingtype",
+      notes: "can post trainingtype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:trainingtype",
+      notes: "can put trainingtype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:trainingtype",
+      notes: "can get trainingtype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:trainingtype",
+      notes: "can delete trainingtype",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:finance",
+      notes: "can post finance",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:finance",
+      notes: "can put finance",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:finance",
+      notes: "can get finance",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:finance",
+      notes: "can delete finance",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:asset",
+      notes: "can post asset",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:asset",
+      notes: "can put asset",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:asset",
+      notes: "can get asset",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:asset",
+      notes: "can delete asset",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:receipt",
+      notes: "can post receipt",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:receipt",
+      notes: "can put receipt",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:receipt",
+      notes: "can get receipt",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:receipt",
+      notes: "can delete receipt",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:remuneration",
+      notes: "can post remuneration",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:remuneration",
+      notes: "can put remuneration",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:remuneration",
+      notes: "can get remuneration",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:remuneration",
+      notes: "can delete remuneration",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:post:payment",
+      notes: "can post payment",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:put:payment",
+      notes: "can put payment",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:get:payment",
+      notes: "can get payment",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+    {
+      name: "can:delete:payment",
+      notes: "can delete payment",
+      createdAt: new Date(),
+      updatedAt: new Date()
+    },
+  ], {}),
+
+  down: (queryInterface) => queryInterface.bulkDelete("Permissions", null, {})
+};
